Turn ScanValueProvider into a @Component, to fix interactive completion
